merge

+++ src/main/java/com/finalterm/alumninetwork/repository/UserRepository.java
@@ -14,5 +14,6 @@ public interface UserRepository {
     User getUserById(Integer userId);
     List<User> getAllUserExactAdmin();
     List<User> getUserByIds(List<Integer> userIds);
-    long countUsers();
+    int countUsers();
 }
+
